Add email preparation function to 'prepare_mail.py'
The 'run' function was added to a new 'prepare_mail.py' file. This function prepares and sends an email using environment variables ("main_mail", "store_mail" and "subject") for recipients and server details. The email body includes a HTML content body and an attachment file. This functionality was added to automate the process of sending email reports to specific recipients.