Refactor variable names and disable price comparison
Variable names related to the date lists have been refactored for 'GRESKO' and 'KALIMERA FROUTA' to improve the clarity and readability of the code. Also, the functionality that compares the old and new prices for both has been temporarily disabled, probably due to some issues that need to be resolved, as evident from these lines being commented out. The further step of codebase optimization could be resetting these processes after troubleshooting and debugging.