Increase width for Excel columns, adjustments for autifit
The width of columns A, B, E, and F has been increased for the sake of improved readability in the generated Excel document. The autofit function has been commented out, and manual sizing has been applied instead. Columns C and D have been introduced as well with a width of 1. The changes have been made based on the need to improve the design and usability of the reports for end users.