Update list_of_dates and uncomment myfunc.run
Added a new date '231107' to the list_of_dates_kalimera and uncommented the function myfunc.run to allow the program to send emails. The addition of the date will ensure that all relevant data will be considered in the process, and the uncommented function will enable the email sending feature.