Enable price drop analysis in the main process
The commented out sections in the 'main.py' script that handle price drop analysis for 'Gresco' and 'Kalimera' are now activated. This change reintroduces the function that reads old price data, compares it with new data and computes price change. This was originally ignored presumably to avoid dealing with price drops, but has been activated now to present a complete picture of price changes. Non-essential print statements continue to be ignored and email execution remains commented out for test runs.